refactor: input folder => inputs folder
