+++ dashboard/src/components/inputs/search-input/search-input.scss
